pyproject.toml
[project]
name = "tg-monitor"
version = "0.1.0"
description = "Telegram 群聊监控 & 智能汇总系统"
requires-python = ">=3.9"
dependencies = [
    "telethon>=1.36",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "click>=8.1",
    "rich>=13.7",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
tg-monitor = "src.cli:cli"

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"
